+++ cognite/neat/rules/exporters/_rules2dms.py
@@ -13,7 +13,7 @@ from cognite.neat.rules.models._rules.dms_schema import DMSSchema
 from cognite.neat.utils.cdf_loaders import ContainerLoader, DataModelingLoader, DataModelLoader, SpaceLoader, ViewLoader
 
 from ._base import BaseExporter
-from ._data_classes import UploadResult
+from ._models import UploadResult
 
 
 class DMSExporter(BaseExporter[DMSSchema]):
